cloud_functions/load_to_bq/requirements.txt
google-cloud-bigquery==1.19.0
google-cloud-storage==1.14.0
pyyaml